extract plot routines
